refactor: 코드 구조 개선 및 모듈화
- 기능별로 디렉토리 구조 재구성
  - dom: DOM 요소 조작 및 참조 관련 코드
  - features: 상품과 금액 관련 핵심 기능
  - store: 전역 상태 관리
  - utils: 포맷팅, 로깅 등 유틸리티 함수

- 관련 기능들을 단일 책임 원칙에 따라 모듈화
- 파일 경로 및 import 구문 정리
